% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Title}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{A factor}

\item{b}{A factor}
}
\value{
A factor with combined levels from a and b
}
\description{
Title
}
\examples{
fbind(as.factor(c("dog", "cat")), as.factor(c("gerbil", "parakeet")))
}
