% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage22F001-package.R
\docType{package}
\name{testpackage22F001-package}
\alias{testpackage22F001}
\alias{testpackage22F001-package}
\title{testpackage22F001: Test case for 413-613}
\description{
demonstrate the ideas of building a packge, .
}
\author{
\strong{Maintainer}: Richard Ressler \email{rressler@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
